feat [Alked Tang] #22 Included timezone selection in schedule view. - Moved timezones.js from model/edit to model/data because of its current general purpose.
